tests: api: separate live tests
